backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
pydantic
sentence-transformers
python-multipart
sqlalchemy
mysql-connector-python
python-dotenv
deep-translator
langdetect
indic-transliteration
# Speech-to-text (offline)
vosk
soundfile
faster-whisper
openai-whisper
scikit-learn
wordfreq
transformers
sentencepiece
torch
keybert